requirements.txt
fastapi
uvicorn
google-genai
pydantic
pillow
python-multipart
groq
